EmotionCardをHomeViewから分離
